Remove wildcard imports (`import *`)
Wildcard imports are nice and succinct, but they make code less
readable, since it is not clear where objects are defined. If you want
to know more about `Monster`, you need some indicator of which file
defines the class. PEP 8 explicitly says to avoid wildcard imports.

I considered an alternate pattern of explicitly importing each used
class, like:
    from api.models import (CharClass, Document, <fifteen more>)
But I decided that that was so long that it was unwieldy.

I tested this change by importing everything and getting every endpoint.