Create document.json
This is a weird one because the book is no longer being sold directly by the original publisher. I opted to include a link to a blog post from the publisher about the book, which contains a link to where the product used to be listed on the store; the link currently redirects to Green Ronin's other 5e products. You can find the original store link on archive.org, for instance: https://web.archive.org/web/20191230102305/https://greenroninstore.com/collections/fifth-edition/products/critical-role-tal-dorei-campaign-setting-5e-pdf

Under OGL Lines, I did not reproduce the entire OGL license in full (should I?), but I did include the product specific paragraphs from the license page: the copyright notice, the declaration of Product Identity, and the declaration of Open Content.